% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_api.R
\name{got_api}
\alias{got_api}
\title{Retrieves books, characters or houses information from the anapioficeandfire API}
\usage{
got_api(type = c("books", "characters", "houses"), id = NULL, url = NULL,
  ...)
}
\arguments{
\item{type}{The endpoint type, one of books, characters or houses.
Default to books if nothing is provided.}

\item{id}{The object index. Default is all unless an id is specified.}

\item{...}{Additional parameters to pass on to \code{GET} such as query, path, etc.
User can also filter by the parameters of interest like name of a character.}
}
\value{
List with all parameters of that endpoint
}
\description{
A book is a book that are already published or planned for future.
A Character is an individual within the Ice And Fire universe.
A House is a house branch within the Ice And Fire universe.
}
\examples{
\dontrun{

# Retrieve books on page 1
books_page1 <- got_api(type = "books")

# Retrieve books on page 1, change page size to 20. Default is 10.
booksAll <- got_api(type = "books", query = list(pageSize="20"))

# Retrieve characters on page 3, change page size to 20.
characters_page_3 <- got_api(type = "characters", query = list(page = "3", pageSize="20"))

# Retrieve character Jon Snow by id
jon_snow <- got_api(type = "characters", id = "583")

# Retrieve character Jon Snow by name
jon_snow_by_name <- got_api(type = "characters", query = list(name = "Jon Snow"))

# Retrieve house id 378
house378 <- got_api(type = "houses", id = 378)
}
}
\author{
Mango Solutions
}
