% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoTr-package.R
\docType{package}
\name{Gotr-package}
\alias{Gotr-package}
\alias{Gotr}
\title{R wrapper for An API of Ice And Fire}
\description{
This package provides access to the \href{https://anapioficeandfire.com/}{anapioficeandfire}
API from R. It pulls data of the universe of Ice and Fire including HBO series Game of Thrones.
}
